Add pagination support to client fetch people methods
Update the internal base task completion-handler
method so the completion generated also passes in
the full JSON object. This provides the
flexibility for the internal base task methods to
make use of the non-results JSON data (i.e.
pagination) as needed. This also requires updating
each base task method.

Next update the base fetch task method and the
fetch people methods to allow an optional
reference to a pagination object used by the app
to be passed. If passed, the object referenced
will be updated based on the newly passed full
JSON object on completion.