Add collection view controller and cell classes that implement protocol
Our view controllers and views should manage a
data source and be able to work with custom nib
names for them and their child views.

Transitions follow the style of storyboard segues,
but since we don't have a storyboard, we're just
calling an action method when performing.

Includes pull-based action additions heavily
inspired by SVPullToRefresh. Add the labels (with
some elbow grease) to the collection view as
decoration views.

While section headers resembles the table view
section headers, the ability for the header to
stick to below the navigation bar is left as an
exercise (hint: subclass flow layout).