Initial draft of accounts view controller
The view controller observes its data source to
respond to state changes. It is mostly a
conventional view controller (UIAppearance,
customizable attributes, etc.) with an account
view and action buttons that only shows as needed
and update themselves depending on state.

By default, the view controller automatically
prompts for the nation slug if there is no active
account.