Initial accounts picker and flow for multiple accounts
Includes adding checks to other toggle subview
methods to only animate when needed. This is
needed because the picker view glitches when
redundant animations occur.

Also adds an error alert noting (for now) multiple
signed-in accounts per nation isn't supported.
This message is long, so alert auto-dismissal is
just removed. The flow is kept by storing the
error alert.

Note the picker rows are customized using
attributed strings. Also note the picker view
needs to be contained to be collapsible.