Add major refinements and expansions to account button
'*Only' button types allow rendering the button in
limited spaces. The default shows both the avatar
and name. If the desired button type is unsuitable
given provided data, a fallback, 'actual' button
type is set.

The name label now also doubles for showing icons
in the new icon-only type. The icon depends on the
provided data. Also adds a flag for if the context
has multiple active accounts.

Also adds an avatar-only type, and comments in the
sample app integration to explain the
possibilities.

Also, corner radius is now customizable, there is
a new flag for circular avatars, and the avatar
now has a tinted border and tweaked size.