Add support for persisting same-nation accounts
This requires updating the entire accounts and
auth aspect:

- The authenticator needs to allow its credential
identifier to be mutable.

- The account needs to be able to customize the
credential identifier to also contain the user
name and id, so it will be saving the credential
manually and turning off its authenticator's
auto-save.

- The accounts manager needs include the user id
when persisting and restoring accounts. It also
should prevent activating duplicates and try to
always keep an account active.

Includes minor refactoring and reorganization.