Add support for account persistence for n managers
This requires a dynamic yet nonrandom identifier.
The default implementation, using the delegate's
class name, only works partially (as noted). But
it also requires the manager's delegate be a
required property, which consequently updates a
lot of code. Flags and options have also been
changed to delegate methods as needed.

The real solution is to allow the delegate to
customize the identifier and ensure the
user-defaults key's uniqueness.