Upgrade async testing in test case to Xcode 6
Xcode 6 provides test expectations. Our original
interface is simpler, so it's kept. It's also
modified to be more convenient for a test to opt
into HTTP stubbing.

The other benefit is things like testing code that
uses dispatch_async are no longer a problem now
that we're not manually manipulating run loops.