Always sign out when activating first account
The accounts manager cannot assume that it is the
only manager affecting which NationBuilder session
Safari currently uses. 

While this is always true, we can be lenient and
require no prior signout if the nation is
different, ie. a use case in which a user has only
one account for each of several nations, instead
of a use case in which a user has one nation with
several accounts.